requirements.txt
asttokens==3.0.1
cffi==2.0.0
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
dotenv==0.9.9
executing==2.2.1
fonttools==4.61.1
ImageIO==2.37.2
ipykernel==7.2.0
ipympl==0.10.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lazy_loader==0.4
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.2
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.1
parso==0.8.6
patsy==1.0.2
pillow==12.1.1
platformdirs==4.9.2
prompt_toolkit==3.0.52
psutil==7.2.2
pure_eval==0.2.3
pycparser==3.0
Pygments==2.19.2
pyparsing==3.3.2
pyspnego==0.12.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyzmq==27.1.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
six==1.17.0
smbprotocol==1.16.0
spectral==0.24
sspilib==0.5.0
stack-data==0.6.3
statsmodels==0.14.6
threadpoolctl==3.6.0
tifffile==2026.2.24
tornado==6.5.4
traitlets==5.14.3
tzdata==2025.3
wcwidth==0.6.0
webcolors==25.10.0
widgetsnbextension==4.0.15
xgboost==3.2.0
